Break up DB object into package.
