+++ src/main/scala/com/ironcorelabs/davenport/db/DBOp.scala
@@ -1,0 +1,15 @@
+//
+// Copyright (c) 2015 IronCore Labs
+//
+package com.ironcorelabs.davenport.db
+
+import scalaz.\/
+
+/** Any database operation must be represented by a `DBOp` */
+sealed trait DBOp[A]
+final case class GetDoc(key: Key) extends DBOp[DBError \/ DBValue]
+final case class CreateDoc(key: Key, doc: RawJsonString) extends DBOp[DBError \/ DBValue]
+final case class UpdateDoc(key: Key, doc: RawJsonString, commitVersion: CommitVersion) extends DBOp[DBError \/ DBValue]
+final case class RemoveKey(key: Key) extends DBOp[DBError \/ Unit]
+final case class GetCounter(key: Key) extends DBOp[DBError \/ Long]
+final case class IncrementCounter(key: Key, delta: Long = 1) extends DBOp[DBError \/ Long]
