+++ src/main/scala/com/ironcorelabs/davenport/syntax/key.scala
@@ -4,7 +4,7 @@
 package com.ironcorelabs.davenport
 package syntax
 
-import DB.{ Key, DBProg }
+import db.{ Key, DBProg, DBDocument }
 import argonaut.{ DecodeJson, EncodeJson, CodecJson }
 
 // The convention is for syntax objects to start with lower case, so they look
@@ -16,8 +16,8 @@ trait KeyOps {
     def dbGet[T](implicit codec: DecodeJson[T]): DBProg[DBDocument[T]] = DBDocument.get(key)(codec)
     def dbRemove: DBProg[Unit] = DBDocument.remove(key)
     def dbCreate[T](t: T)(implicit codec: EncodeJson[T]): DBProg[DBDocument[T]] = DBDocument.create(key, t)
-    def dbIncrementCounter(delta: Long): DBProg[Long] = DB.incrementCounter(key, delta)
-    def dbGetCounter: DBProg[Long] = DB.getCounter(key)
+    def dbIncrementCounter(delta: Long): DBProg[Long] = db.incrementCounter(key, delta)
+    def dbGetCounter: DBProg[Long] = db.getCounter(key)
     def dbModify[T](f: T => T)(implicit codec: CodecJson[T]): DBProg[DBDocument[T]] = DBDocument.modify(key, f)
   }
 }
